# zarf/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sample implementation of an API using the carbon framework",
        "title": "Carbon Sample API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/api/v0",
    "paths": {
        "/task": {
            "post": {
                "description": "Handles the live cycle of tasks: CRUD + list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "create a new task",
                "parameters": [
                    {
                        "description": "user login payload",
                        "name": "UserData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlrs.loginData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlrs.userStatus"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "Handles the user login process",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Login a user",
                "parameters": [
                    {
                        "description": "user login payload",
                        "name": "UserData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlrs.loginData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlrs.userStatus"
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "get": {
                "description": "Logs out the current user based on the session cookie",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Logout a user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlrs.userStatus"
                        }
                    }
                }
            },
            "put": {
                "description": "Logs out the current user based on the session cookie",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Logout a user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlrs.userStatus"
                        }
                    }
                }
            },
            "post": {
                "description": "Logs out the current user based on the session cookie",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Logout a user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlrs.userStatus"
                        }
                    }
                }
            }
        },
        "/user/options": {
            "get": {
                "description": "Get options specific to the currently logged-in user based on the session cookie",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user options",
                "responses": {
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/middleware.jsonErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.jsonErr"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/middleware.jsonErr"
                        }
                    }
                }
            }
        },
        "/user/status": {
            "get": {
                "description": "Show the satus information about the current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlrs.userStatus"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlrs.loginData": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "handlrs.userStatus": {
            "type": "object",
            "properties": {
                "logged-req": {
                    "type": "boolean"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "middleware.jsonErr": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}